Add files via upload
For any organization, even the slightest chance of financial risk can not be ignored. Organizations conduct regular inspections on their expenditures and revenue to make sure that they do not fall below the critical limit. In this hackathon, you as a data scientist must use the given data to predict whether an organization is under a possible financial risk or not.

Given are 7 distinguishing factors that can provide insight into whether an organization may face a financial risk or not. Your objective as a data scientist is to build a machine learning model that can predict if an organization will fall under the risk using the given features.

Data Description
The unzipped folder will have the following files.

Train.csv – 543 observations.
Test.csv – 233 observations.
Sample Submission – Sample format for the submission.
Target Variable: IsUnderRisk

Got 43rd position on the leaderboard
https://www.machinehack.com/course/financial-risk-prediction-weekend-hackathon-5/leaderboard